Move particle materials into file and fix shader state changes
